Add files via upload
Over a hundred thousand people are dying in road accidents in India every year among those 30% of accidents take place due to roads and buildings which are constructed without following any government regulations. To minimize road accidents we need to identify the buildings which are illegally constructed (without following any government regulations) on the roadside. The aim of this paper is divided into two parts. Firstly, segmenting the roads and buildings from the satellite image and secondly, finding out the illegal construction of roads and buildings by matching the current image with the timed image which is authorized by the government authorities. In order to carry out the objective on a larger scale, an innovative method is evolved which combines a semantic segmentation neural network with residual learning and U-net. In addition, it is proposed a method for road and building extraction from a satellite image. The convolution neural network is built using the U-net architecture. The two 
main advantages of this network are (1) U-net is symmetric and (2) the skip connections between the down-sampling path and the up-sampling apply a concatenation operation instead of an addition operation when compared to standard Fully Conventional Neural network (FCN)-8. In this application, we used the Massachusetts Benchmark dataset for roads and building extraction. Essentially, in this research, a comparison was carried out with other methods recently available (RESNET, SEGNET), this method greater advantage in providing results even with not fully connected layers. Moreover, this network was found to be equally good in terms of computational time.
Key terms: Road extraction, Building extraction, Convolutional Neural Network, U-Net, Convolutional 
layer, Max pooling layer, Epoch.